fix: robustness LOWs from the final audit (coords, fz_try, recursion, truncation, race doc)
- DjVu link rects: coords[4] is zero-initialized, so a short
  (rect x y) annotation yields a zero-area link instead of reading
  uninitialized stack into link geometry.
- fz_register_document_handlers moves inside fz_try in the PDF render
  instance loop and the CBR open path; a throw outside fz_try aborts
  the process, and registration now reports through the existing
  error paths instead (OOM-only in practice, but the abort was
  gratuitous).
- The three DjVu miniexp walkers bound their recursion at 64 levels;
  the s-expressions are file-controlled and hostile nesting depth
  could exhaust the stack.
- The MOBI text-record index is guint32: on a combo file with a high
  KF8 boundary, kf8_start + i can exceed 16 bits and the truncated
  index could alias an earlier record past the bound check.
- The CBR unlocked page-size reads are documented as a deliberate
  tolerance (page-count-stable arrays, aligned 8-byte doubles, stale
  by at most one probe write) rather than left looking like an
  oversight.

+++ src/fw-mobi-parser.c
@@ -454,13 +454,13 @@ walk_exth (const guchar *exth, gsize exth_len,
   }
 }
 
-/* ── Image decoding — port of MOBI.loadResource ───────────────
+/* ── Image filtering — port of MOBI.loadResource ──────────────
  *
  * Foliate's loadResource peels off FONT/VIDE/AUDI prefixes and
  * returns raw bytes. For images, the bytes are JPEG/PNG/GIF
- * directly — gdk_texture_new_from_bytes handles all three.
- * Magic-byte check filters obviously-non-image records (FLIS,
- * FCIS, BOUNDARY, etc.) before we hand bytes to GdkTexture. */
+ * directly — WebKit decodes all three. The magic-byte check
+ * filters obviously-non-image records (FLIS, FCIS, BOUNDARY,
+ * etc.) before they reach the framework-img: table. */
 static gboolean
 looks_like_image (const guchar *p, gsize n)
 {
@@ -1192,7 +1192,10 @@ fw_mobi_parse (const char *path, GError **error)
 
   GString *body = g_string_sized_new (text_records * 4096);
   for (guint16 i = 1; i <= text_records; i++) {
-    guint16 abs_i = (guint16) (kf8_start + i);
+    /* guint32: on a combo file with a high boundary record the sum can
+     * exceed 16 bits, and a truncated abs_i could alias an earlier
+     * record past the bound check. */
+    guint32 abs_i = (guint32) kf8_start + i;
     if (abs_i >= record_count) break;
     gsize off = roff[abs_i];
     gsize len = roff[abs_i + 1] - off;
